Rename tests/test_utils.py to tests/utils.py
